feat: 폴더구조 변경, 더 세부적으로 모듈화 .git/COMMIT_EDITMSG  #79
