--- app/(dashboard)/schedule-dashboard/manage/page.tsx
@@ -1,8 +1,0 @@
-import React from "react";
-import { ScheduleManagePage } from "@/src/views/schedule-manage/index";
-
-const page = () => {
-  return <ScheduleManagePage />;
-};
-
-export default page;
