+++ app/(schedule)/schedule-dashboard/register/page.tsx
@@ -1,0 +1,7 @@
+import { ScheduleRegisterPage } from "@/src/views/schedule/register/index";
+
+const page = () => {
+  return <ScheduleRegisterPage />;
+};
+
+export default page;
