+++ src/entities/schedule/preview-schedule-card/api/get-schedule-preview.ts
@@ -1,0 +1,7 @@
+import { SchedulePreview } from "@/src/entities/schedule/preview-schedule-card/api/types";
+
+export async function getSchedulePreview(): Promise<SchedulePreview[]> {
+  const res = await fetch("/api/schedules");
+  if (!res.ok) throw new Error("Failed to fetch");
+  return res.json();
+}
