+++ src/entities/schedule/preview-schedule-card/api/types.ts
