+++ src/entities/schedule/preview-schedule-card/index.ts
@@ -1,0 +1,1 @@
+export { PreviewScheduleCard } from "@/src/entities/schedule/preview-schedule-card/ui/preview-schedule-card";
