--- src/feature/schedule-manage/index.ts
@@ -1,5 +1,0 @@
-export { ShowScheduleCard } from "../../entities/schedule-card/ui/show-schedule-card";
-export { FilterSchedule } from "./ui/filter-schedule";
-export { CreateSchedule } from "./ui/create-schedule/create-schedule";
-export { SelectOption } from "./ui/select";
-export { ScheduleModal } from "./ui/schedule-modal";
