+++ src/feature/schedule/create-schedule/index.ts
@@ -1,0 +1,1 @@
+export { CreateSchedule } from "@/src/feature/schedule/create-schedule/ui/create-schedule";
