+++ src/feature/schedule/filter-schedule/index.ts
@@ -1,0 +1,1 @@
+export { FilterSchedule } from "@/src/feature/schedule/filter-schedule/ui/filter-schedule";
