+++ src/feature/schedule/filter-schedule/ui/filter-schedule.tsx
