+++ src/feature/schedule/schedule-calender/index.ts
@@ -1,0 +1,1 @@
+export { ScheduleCalender } from "@/src/feature/schedule/schedule-calender/ui/schedule-calender";
