+++ src/feature/schedule/schedule-manage-modal/index.ts
@@ -1,0 +1,1 @@
+export { ScheduleManageModal } from "@/src/feature/schedule/schedule-manage-modal/ui/schedule-manage-modal";
