+++ src/feature/select-option/index.ts
@@ -1,0 +1,1 @@
+export { SelectOption } from "@/src/feature/select-option/ui/select-option";
