--- src/views/schedule-manage/index.ts
@@ -1,2 +1,0 @@
-export { ScheduleRegisterPage } from "./ui/Page/schedule-register";
-export { ScheduleManagePage } from "./ui/Page/schedule-manage";
