+++ src/views/schedule/manage/index.ts
@@ -1,0 +1,1 @@
+export { ScheduleManagePage } from "@/src/views/schedule/manage/ui/page/schedule-manage";
